+++ packages/components/src/Button/ButtonToggle.stories.tsx
@@ -25,7 +25,7 @@
  */
 
 import type { Story } from '@storybook/react/types-6-0'
-import type { Page } from 'puppeteer'
+// import type { Page } from 'puppeteer'
 import React, { useState } from 'react'
 import { defaultArgTypes as argTypes } from '../../../../apps/storybook/src/defaultArgTypes'
 import { Popover } from '../Popover'
@@ -66,12 +66,13 @@ Focused.args = {
 }
 
 Focused.parameters = {
-  beforeScreenshot: async (page: Page) => {
-    const button = await page.$('button')
-    await button?.type(' ')
-    await page.waitForTimeout(200)
-  },
+  // beforeScreenshot: async (page: Page) => {
+  //   const button = await page.$('button')
+  //   await button?.type(' ')
+  //   await page.waitForTimeout(200)
+  // },
   docs: { disable: true },
+  storyshots: { disable: true },
 }
 
 export const InitialValue = Template.bind({})