+++ packages/filter-components/src/Filter/components/AdvancedFilter/components/DateFilter/utils/index.ts
@@ -26,3 +26,4 @@
 export * from './format_date'
 export * from './relative_timeframe_conversions'
 export * from './relative_timeframe_to_string'
+export * from './get_component_locale'